+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Prepend new items to make git merging easier.
 
 ## 0.100.3
 - fix(multi-select): don't set invalid when blurring to click an option from the dropdown
-- feat(multi-autocompleter): allow setting value with types: single id as string or array of ids
+- feat(multi-autocompleter): allow setting value with types array of ids
 - feat(form): allow prop to hide: 'Save & Add Another' button
 
 ## 0.100.1